watcher/requirements.txt
watchdog==4.0.0
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
anthropic==0.40.0
Pillow==11.0.0
